refactor: separate kernel and implement dynamic app system
This major architectural refactoring separates core OS functionalities into a new `window/` (kernel) directory and replaces the static app management system with a dynamic, discovery-based one.

- Kernel Separation: Core UI (`Desktop`, `AppWindow`, etc.) and window management logic (`useWindowManager` hook) are now isolated in the `window/` directory. An `AGENTS.md` file protects this core logic.

- Dynamic App Discovery: A new backend API (`/api/apps`) now discovers applications by scanning for `.app` JSON files in the `components/apps/` directory.

- Frontend Refactoring: The frontend uses a new `AppContext` to consume the dynamic app list. All UI components (`StartMenu`, `Desktop`, `Taskbar`) are refactored to use this new system for display, launching, and pinning.

- "Open With..." Feature: A new "Open with..." context menu option for files dynamically lists all apps that declare file handling capabilities in their `.app` definition.
