+++ components/apps/FileExplorer.app
@@ -1,0 +1,6 @@
+{
+  "name": "File Explorer",
+  "appId": "fileExplorer",
+  "icon": "fileExplorer",
+  "isPinned": true
+}
