+++ components/apps/Settings.app
@@ -1,0 +1,6 @@
+{
+  "name": "Settings",
+  "appId": "settings",
+  "icon": "settings",
+  "isPinned": true
+}
