+++ window/AGENTS.md
@@ -1,0 +1,11 @@
+# Agent Instructions for the /window Directory
+
+## **ATTENTION**
+
+The files within this `window/` directory and its subdirectories constitute the **core kernel** of this simulated operating system. They manage windowing, state, and the main UI shell.
+
+**DO NOT MODIFY** these files unless you are specifically tasked with refactoring or altering a core system behavior.
+
+- **Modifying these files can easily break the entire application.**
+- Treat this directory as a stable, internal API.
+- Application-specific logic belongs in the `/components/apps/` directory.
