+++ window/components/Taskbar.tsx
@@ -1,19 +1,27 @@
-
-import React, { useState, useEffect } from 'react';
-import { OpenApp, AppDefinition } from '../types';
-import { StartIcon, TASKBAR_HEIGHT } from '../constants';
-import { APP_DEFINITIONS } from './apps';
-import { useTheme } from './theme';
-
+import React, { useState, useEffect, useContext, useMemo } from 'react';
+import { OpenApp } from '../../types';
+import { DiscoveredAppDefinition, AppContext } from '../../components/AppContext';
+import { APP_DEFINITIONS } from '../../components/apps';
+import { StartIcon, TASKBAR_HEIGHT, FileGenericIcon } from '../../constants';
+import { useTheme } from '../../components/theme';
 
 interface TaskbarProps {
   openApps: OpenApp[];
   activeAppInstanceId: string | null;
   onToggleStartMenu: () => void;
-  onAppIconClick: (appId: string, instanceId?: string) => void;
+  onAppIconClick: (app: DiscoveredAppDefinition, instanceId?: string) => void;
 }
 
+const getIconComponent = (iconId?: string) => {
+    if (iconId) {
+        const appDef = APP_DEFINITIONS.find(def => def.id === iconId);
+        if (appDef) return appDef.icon;
+    }
+    return FileGenericIcon;
+};
+
 const Taskbar: React.FC<TaskbarProps> = ({ openApps, activeAppInstanceId, onToggleStartMenu, onAppIconClick }) => {
+  const { apps: discoveredApps } = useContext(AppContext);
   const [currentTime, setCurrentTime] = useState(new Date());
   const { theme } = useTheme();
 
@@ -22,22 +30,26 @@ const Taskbar: React.FC<TaskbarProps> = ({ openApps, activeAppInstanceId, onTogg
     return () => clearInterval(timerId);
   }, []);
 
-  const pinnedApps = APP_DEFINITIONS.filter(appDef => appDef.isPinnedToTaskbar);
-  
-  // Combine pinned apps and unique open apps for taskbar display
-  const taskbarItemsMap = new Map<string, {appDef: AppDefinition, instance?: OpenApp}>();
+  const taskbarItems = useMemo(() => {
+    const items = new Map<string, { appDef: DiscoveredAppDefinition, instance?: OpenApp }>();
 
-  pinnedApps.forEach(appDef => {
-    taskbarItemsMap.set(appDef.id, { appDef });
-  });
+    // Add pinned apps first
+    const pinnedApps = discoveredApps.filter(app => app.isPinned);
+    pinnedApps.forEach(appDef => {
+      items.set(appDef.appId || appDef.path!, { appDef });
+    });
 
-  openApps.forEach(openApp => {
-    // If already pinned, update with instance info. Otherwise, add as an open app.
-    const existing = taskbarItemsMap.get(openApp.id);
-    taskbarItemsMap.set(openApp.id, { appDef: existing?.appDef || openApp, instance: openApp });
-  });
-  
-  const taskbarItems = Array.from(taskbarItemsMap.values());
+    // Add or update with open apps
+    openApps.forEach(openApp => {
+      const key = openApp.id; // This is the appId from the original static definition
+      const appDef = discoveredApps.find(app => app.appId === key);
+      if (appDef) {
+        items.set(key, { appDef, instance: openApp });
+      }
+    });
+
+    return Array.from(items.values());
+  }, [discoveredApps, openApps]);
 
 
   return (
@@ -60,17 +72,18 @@ const Taskbar: React.FC<TaskbarProps> = ({ openApps, activeAppInstanceId, onTogg
             const isActive = instance?.instanceId === activeAppInstanceId && !instance?.isMinimized;
             const isMinimized = instance?.isMinimized;
             const isOpen = !!instance;
+            const IconComponent = getIconComponent(appDef.icon);
 
             return (
               <button
                 key={appDef.id}
-                onClick={() => onAppIconClick(appDef.id, instance?.instanceId)}
+                onClick={() => onAppIconClick(appDef, instance?.instanceId)}
                 className={`p-2 rounded h-[calc(100%-8px)] flex items-center relative transition-colors duration-150 ease-in-out
                             ${isActive ? theme.taskbar.activeButton : theme.taskbar.buttonHover}
                             ${isMinimized ? 'opacity-70' : ''}`}
                 title={appDef.name}
               >
-                <appDef.icon className="w-5 h-5" isSmall />
+                <IconComponent className="w-5 h-5" isSmall />
                 {isOpen && (
                   <span className={`absolute bottom-0 left-1/2 transform -translate-x-1/2 w-4 h-1 rounded-t-sm 
                                   ${isActive ? theme.taskbar.activeIndicator : isMinimized ? 'bg-gray-400' : 'bg-gray-500'}`}></span>
@@ -83,10 +96,6 @@ const Taskbar: React.FC<TaskbarProps> = ({ openApps, activeAppInstanceId, onTogg
 
       {/* System Tray - right aligned */}
       <div className="flex items-center space-x-3 text-xs">
-        {/* Placeholder for Wifi, Sound, Battery */}
-        {/* <WifiIcon className="w-4 h-4" />
-        <SoundIcon className="w-4 h-4" />
-        <BatteryIcon className="w-4 h-4" /> */}
         <div>{currentTime.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}</div>
         <div>{currentTime.toLocaleDateString([], { month: 'short', day: 'numeric' })}</div>
       </div>
